Updates VM connection impl (makes it easier to test).
Changes underyling FlutterFuchsia implementation to use a Driver class
similar to Flutter's existing driver library.
